--- index.d.ts
@@ -1,13 +1,0 @@
-import ReactNativeZoomableView from './src/ReactNativeZoomableView';
-import ReactNativeZoomableViewWithGestures from './src/ReactNativeZoomableViewWithGestures';
-import {
-  ReactNativeZoomableViewProps,
-  ZoomableViewEvent,
-} from '@openspacelabs/react-native-zoomable-view';
-
-export {
-  ReactNativeZoomableView,
-  ReactNativeZoomableViewWithGestures,
-  ReactNativeZoomableViewProps,
-  ZoomableViewEvent,
-};
